split awc response module
